Use common Application class for library COM apps
